parse text sizes as floats

+++ src/text_placements.cpp
@@ -168,7 +168,7 @@ void text_placements_simple::set_positions(std::string positions)
 
     std::string::iterator first = positions.begin(),  last = positions.end();
     qi::phrase_parse(first, last,
-		     (direction_name[push_back(phoenix::ref(direction_), _1)] % ',') >> *(',' >> qi::int_[push_back(phoenix::ref(text_sizes_), _1)]),
+		     (direction_name[push_back(phoenix::ref(direction_), _1)] % ',') >> *(',' >> qi::float_[push_back(phoenix::ref(text_sizes_), _1)]),
 		     space
     );
     if (first != last) {
